refactor(json): extract format errors to module
